feat: add onboarding billing step (#17514)
* update billing hero

* restructure plan comparison so it can be used w/o modal

* no onboarding for inclusion-only products

* add basic billing step

* improve variable naming

* fix

* make billing step work after subscribe & addons

* automatically include billing step in onboarding

* show billing step properly

* tolerate sub'd product on intro

* store when someone has sub'd during onboarding

* just put addons on the main screen
