--- frontend/src/scenes/billing/PlanComparisonModal.scss
@@ -1,42 +1,0 @@
-.PlanComparisonModal {
-    max-width: 900px;
-    table {
-        table-layout: fixed;
-        td {
-            vertical-align: top;
-            padding: 0.75rem 1.25rem;
-
-            &.PlanTable__td__upgradeButton {
-                padding-top: 1rem;
-                padding-bottom: 1rem;
-            }
-        }
-        th {
-            vertical-align: top;
-            padding: 0.75rem 1.25rem 0.75rem;
-            font-weight: 600;
-            text-align: left;
-
-            &.PlanTable__th__section {
-                padding: 0.25rem 1.25rem 0.25rem;
-                font-weight: 500;
-            }
-
-            &.PlanTable__th__feature {
-                padding: 0.75rem 1.25rem 0.75rem 3.25rem;
-                font-weight: 600;
-            }
-
-            &.PlanTable__th__last-feature {
-                padding-bottom: 2rem;
-            }
-
-            p {
-                font-weight: 400;
-            }
-        }
-    }
-    .PlanTable__tr__border {
-        border-bottom: 3px rgba(0, 0, 0, 0.07) dotted;
-    }
-}
