+++ frontend/src/scenes/onboarding/OnboardingStep.tsx
@@ -8,13 +8,17 @@ export const OnboardingStep = ({
     title,
     subtitle,
     children,
+    showSkip = false,
+    continueOverride,
 }: {
     title: string
     subtitle?: string
     children: React.ReactNode
+    showSkip?: boolean
+    continueOverride?: JSX.Element
 }): JSX.Element => {
-    const { onboardingStep, totalOnboardingSteps } = useValues(onboardingLogic)
-    const { incrementOnboardingStep, decrementOnboardingStep, completeOnboarding } = useActions(onboardingLogic)
+    const { currentOnboardingStepNumber, totalOnboardingSteps } = useValues(onboardingLogic)
+    const { setCurrentOnboardingStepNumber, completeOnboarding } = useActions(onboardingLogic)
     return (
         <BridgePage
             view="onboarding-step"
@@ -22,29 +26,51 @@ export const OnboardingStep = ({
             hedgehog={false}
             fixedWidth={false}
             header={
-                onboardingStep > 1 && (
+                currentOnboardingStepNumber > 1 && (
                     <div className="mb-4">
-                        <LemonButton icon={<IconArrowLeft />} onClick={decrementOnboardingStep}>
+                        <LemonButton
+                            icon={<IconArrowLeft />}
+                            onClick={() => setCurrentOnboardingStepNumber(currentOnboardingStepNumber - 1)}
+                        >
                             Back
                         </LemonButton>
                     </div>
                 )
             }
         >
-            <div className="max-w-md">
+            <div className="w-md">
                 <h1 className="font-bold">{title}</h1>
                 <p>{subtitle}</p>
                 {children}
                 <div className="mt-8 flex justify-end gap-x-2">
-                    <LemonButton
-                        type="primary"
-                        onClick={() =>
-                            onboardingStep == totalOnboardingSteps ? completeOnboarding() : incrementOnboardingStep()
-                        }
-                        sideIcon={onboardingStep !== totalOnboardingSteps ? <IconArrowRight /> : null}
-                    >
-                        {onboardingStep == totalOnboardingSteps ? 'Finish' : 'Continue'}
-                    </LemonButton>
+                    {showSkip && (
+                        <LemonButton
+                            type="tertiary"
+                            onClick={() =>
+                                currentOnboardingStepNumber == totalOnboardingSteps
+                                    ? completeOnboarding()
+                                    : setCurrentOnboardingStepNumber(currentOnboardingStepNumber + 1)
+                            }
+                            status="muted"
+                        >
+                            Skip for now
+                        </LemonButton>
+                    )}
+                    {continueOverride ? (
+                        continueOverride
+                    ) : (
+                        <LemonButton
+                            type="primary"
+                            onClick={() =>
+                                currentOnboardingStepNumber == totalOnboardingSteps
+                                    ? completeOnboarding()
+                                    : setCurrentOnboardingStepNumber(currentOnboardingStepNumber + 1)
+                            }
+                            sideIcon={currentOnboardingStepNumber !== totalOnboardingSteps ? <IconArrowRight /> : null}
+                        >
+                            {currentOnboardingStepNumber == totalOnboardingSteps ? 'Finish' : 'Continue'}
+                        </LemonButton>
+                    )}
                 </div>
             </div>
         </BridgePage>
