+++ frontend/src/scenes/onboarding/onboardingLogic.tsx
@@ -8,6 +8,7 @@ import { billingLogic } from 'scenes/billing/billingLogic'
 export interface OnboardingLogicProps {
     productKey: ProductKey | null
 }
+export type AllOnboardingSteps = JSX.Element[]
 
 export const onboardingLogic = kea<onboardingLogicType>({
     props: {} as OnboardingLogicProps,
@@ -19,11 +20,11 @@ export const onboardingLogic = kea<onboardingLogicType>({
     actions: {
         setProduct: (product: BillingProductV2Type | null) => ({ product }),
         setProductKey: (productKey: string | null) => ({ productKey }),
-        setOnboardingStep: (onboardingStep: number) => ({ onboardingStep }),
-        incrementOnboardingStep: true,
-        decrementOnboardingStep: true,
-        setTotalOnboardingSteps: (totalOnboardingSteps: number) => ({ totalOnboardingSteps }),
+        setCurrentOnboardingStepNumber: (currentOnboardingStepNumber: number) => ({ currentOnboardingStepNumber }),
         completeOnboarding: true,
+        setAllOnboardingSteps: (allOnboardingSteps: AllOnboardingSteps) => ({ allOnboardingSteps }),
+        setStepKey: (stepKey: string) => ({ stepKey }),
+        setSubscribedDuringOnboarding: (subscribedDuringOnboarding) => ({ subscribedDuringOnboarding }),
     },
     reducers: () => ({
         productKey: [
@@ -38,18 +39,16 @@ export const onboardingLogic = kea<onboardingLogicType>({
                 setProduct: (_, { product }) => product,
             },
         ],
-        onboardingStep: [
+        currentOnboardingStepNumber: [
             1,
             {
-                setOnboardingStep: (_, { onboardingStep }) => onboardingStep,
-                incrementOnboardingStep: (state) => state + 1,
-                decrementOnboardingStep: (state) => state - 1,
+                setCurrentOnboardingStepNumber: (_, { currentOnboardingStepNumber }) => currentOnboardingStepNumber,
             },
         ],
-        totalOnboardingSteps: [
-            1,
+        allOnboardingSteps: [
+            [] as AllOnboardingSteps,
             {
-                setTotalOnboardingSteps: (_, { totalOnboardingSteps }) => totalOnboardingSteps,
+                setAllOnboardingSteps: (_, { allOnboardingSteps }) => allOnboardingSteps as AllOnboardingSteps,
             },
         ],
         onCompleteOnbardingRedirectUrl: [
@@ -69,7 +68,26 @@ export const onboardingLogic = kea<onboardingLogicType>({
                 },
             },
         ],
+        subscribedDuringOnboarding: [
+            false,
+            {
+                setSubscribedDuringOnboarding: (_, { subscribedDuringOnboarding }) => subscribedDuringOnboarding,
+            },
+        ],
     }),
+    selectors: {
+        totalOnboardingSteps: [
+            (s) => [s.allOnboardingSteps],
+            (allOnboardingSteps: AllOnboardingSteps) => allOnboardingSteps.length,
+        ],
+        shouldShowBillingStep: [
+            (s) => [s.product, s.subscribedDuringOnboarding],
+            (product: BillingProductV2Type | null, subscribedDuringOnboarding: boolean) => {
+                const hasAllAddons = product?.addons?.every((addon) => addon.subscribed)
+                return !product?.subscribed || !hasAllAddons || subscribedDuringOnboarding
+            },
+        ],
+    },
     listeners: ({ actions, values }) => ({
         loadBillingSuccess: () => {
             actions.setProduct(values.billing?.products.find((p) => p.type === values.productKey) || null)
@@ -94,13 +112,16 @@ export const onboardingLogic = kea<onboardingLogicType>({
         },
     }),
     urlToAction: ({ actions }) => ({
-        '/onboarding/:productKey': ({ productKey }) => {
+        '/onboarding/:productKey': ({ productKey }, { success, upgraded }) => {
             if (!productKey) {
                 window.location.href = urls.default()
                 return
             }
+            if (success || upgraded) {
+                actions.setSubscribedDuringOnboarding(true)
+            }
             actions.setProductKey(productKey)
-            actions.setOnboardingStep(1)
+            actions.setCurrentOnboardingStepNumber(1)
         },
     }),
 })
